requirements.txt
# --- requirements.txt ---
fastapi
uvicorn
pillow
torch
torchvision
timm